requirements.txt
Flask>=2.2.2
python-dotenv==0.19.0
langchain
openai
